Change float to double type

+++ include/illu/illu_raw_data.hpp
@@ -6,18 +6,18 @@
 namespace illu {
 
 struct Pos {
-  float x_, y_, z_;
+  double x_, y_, z_;
 };
 
 struct RGB {
-  float b_, g_, r_;
+  double b_, g_, r_;
 };
 
 struct PhotonRecord {
   char reflection_;
   char refraction_;
   RGB rgb_;
-  float depth_;
+  double depth_;
   Pos pos_;
 };
 
@@ -26,7 +26,7 @@ struct PixelRecord {
   char refraction_;
   RGB BRDF_;
   Pos pos_;
-  float depth_;
+  double depth_;
   std::vector<int> photons;
 };
